ISSUES-4006 move MySQL command to the PacketsGeneric.h #4006
